web/requirements.txt
click==6.6
coreapi==2.0.8
coverage==4.2
Django==1.9.10
django-braces==1.9.0
django-cors-headers==1.1.0
django-debug-toolbar==1.5
django-extensions==1.7.4
django-filter==0.15.0
django-jenkins==0.110.0
django-rest-swagger==2.0.5
django-shotgun==0.2
djangorestframework==3.4.7
djangorestframework-gis==0.10.1
drf-extensions==0.2.8
drf-hal-json==0.9.1
drf-nested-fields==0.9.3
elasticsearch==2.4.0
elasticsearch-dsl==2.1.0
factory-boy==2.7.0
fake-factory==0.7.2
first==2.0.1
itypes==1.1.0
Markdown==2.6.7
netifaces==0.10.5
openapi-codec==1.1.0
pip-review==0.4
pip-tools==1.7.0
psycopg2==2.6.2
pyslack-real==0.6.0
python-dateutil==2.5.3
PyYAML==3.12
requests==2.11.1
simplejson==3.8.2
six==1.10.0
sphinx-me==0.3
sqlparse==0.2.1
uritemplate==3.0.0
urllib3==1.18
